+++ .github/workflows/build.yml
@@ -45,3 +45,17 @@ jobs:
 
       - name: Run tests
         run: npm test
+
+      - name: "Archive unit test debug log"
+        if: ${{ failure() }}
+        uses: actions/upload-artifact@v3
+        with:
+          name: ${{ matrix.node-version }}-unit-test-debug.log
+          path: test/temp/debug.log
+
+      - name: "Archive scenario test debug log"
+        if: ${{ failure() }}
+        uses: actions/upload-artifact@v3
+        with:
+          name: ${{ matrix.node-version }}-cucumber-debug.log
+          path: test/temp/debugc.log
