+++ fabric-common/lib/DiscoveryService.js
@@ -5,9 +5,8 @@
  */
 
 const TYPE = 'DiscoveryService';
-const Long = require('long');
 
-const {byteToNormalizedPEM, checkParameter, getLogger} = require('./Utils.js');
+const {checkParameter, getLogger} = require('./Utils.js');
 const ServiceAction = require('./ServiceAction.js');
 const DiscoveryHandler = require('./DiscoveryHandler.js');
 
@@ -15,6 +14,7 @@ const logger = getLogger(TYPE);
 
 const fabproto6 = require('fabric-protos');
 const {SYSTEMCHAINCODES} = require('./Endorser.js');
+const DiscoveryResultsProcessor = require('./DiscoveryResultsProcessor.js');
 
 /**
  * The DiscoveryService class represents a peer in the target fabric network that
@@ -42,6 +42,7 @@ class DiscoveryService extends ServiceAction {
 		this.client = channel.client;
 		this.type = TYPE;
 		this.refreshAge = 5 * 60 * 1000; // 5 minutes default
+		this.refreshRunning = false;
 
 		this.discoveryResults = null;
 		this.asLocalhost = false;
@@ -328,47 +329,19 @@ class DiscoveryService extends ServiceAction {
 		}
 
 		// -----
-		this.discoveryResults = {};
 		logger.debug(`${method} - processing discovery response`);
-		if (response && response.results) {
-			let errorMsg = null;
-			logger.debug(`${method} - parse discovery response.results`);
-			for (const index in response.results) {
-				const result = response.results[index];
-				if (result.result === 'error') {
-					logger.error(`${method} - Channel:${this.channel.name} received discovery error:${result.error.content}`);
-					errorMsg = result.error.content;
-					break;
-				} else {
-					logger.debug(`${method} - process result index:${index}`);
-					if (result.config_result) {
-						logger.debug(`${method} - process result - have configResult in ${index}`);
-						const config = this._processConfig(result.config_result);
-						this.discoveryResults.msps = config.msps;
-						this.discoveryResults.orderers = await this._buildOrderers(config.orderers);
-					}
-					if (result.members) {
-						logger.debug(`${method} - process result - have members in ${index}`);
-						this.discoveryResults.peers_by_org = await this._processMembership(result.members);
-					}
-					if (result.cc_query_res) {
-						logger.debug(`${method} - process result - have ccQueryRes in ${index}`);
-						this.discoveryResults.endorsement_plan = await this._processChaincode(result.cc_query_res);
-					}
-					logger.debug(`${method} - completed processing result ${index}`);
-				}
-			}
-
-			if (errorMsg) {
-				throw Error(`DiscoveryService: ${this.name} error: ${errorMsg}`);
-			} else {
-				this.discoveryResults.timestamp = (new Date()).getTime();
-				return this.discoveryResults;
-			}
-		} else {
+		if (!response || !response.results) {
 			logger.error('%s - no discovery results', method);
 			throw new Error('DiscoveryService has failed to return results');
 		}
+
+		logger.debug(`${method} - parse discovery response.results`);
+
+		const processor = new DiscoveryResultsProcessor(this, response.results);
+		const results = await processor.parseDiscoveryResults();
+		results.timestamp = currentTimestamp();
+		this.discoveryResults = results;
+		return results;
 	}
 
 	/**
@@ -381,24 +354,29 @@ class DiscoveryService extends ServiceAction {
 	async getDiscoveryResults(refresh) {
 		const method = `getDiscoveryResults[${this.name}]`;
 		logger.debug(`${method} - start`);
-		if (!this.discoveryResults && !this.savedResults) {
+
+		if (!this.discoveryResults) {
 			throw Error('No discovery results found');
 		}
-		// when savedResults exist, then a refresh is running
-		if (this.savedResults) {
-			logger.debug(`${method} - using the saved results`);
-			return this.savedResults;
-		}
 
-		if (refresh && (new Date()).getTime() - this.discoveryResults.timestamp > this.refreshAge) {
+		if (refresh && !this.refreshRunning && this._isRefreshRequired()) {
 			logger.debug(`${method} - will refresh`);
-			this.savedResults = this.discoveryResults;
-			await this.send({asLocalhost: this.asLocalhost, requestTimeout: this.requestTimeout, targets: this.targets});
-			this.savedResults = null;
+			this.refreshRunning = true;
+			try {
+				await this.send({asLocalhost: this.asLocalhost, requestTimeout: this.requestTimeout, targets: this.targets});
+			} finally {
+				this.refreshRunning = false;
+			}
 		} else {
 			logger.debug(`${method} - not refreshing`);
 		}
+
 		return this.discoveryResults;
+	}
+
+	_isRefreshRequired() {
+		const resultsAge = currentTimestamp() - this.discoveryResults.timestamp;
+		return resultsAge > this.refreshAge;
 	}
 
 	/**
@@ -408,11 +386,34 @@ class DiscoveryService extends ServiceAction {
 		const method = `hasDiscoveryResults[${this.name}]`;
 		logger.debug(`${method} - start`);
 
-		if (this.discoveryResults) {
-			return true;
+		return !!this.discoveryResults;
+	}
+
+	_buildUrl(hostname = checkParameter('hostname'), port = checkParameter('port')) {
+		const method = `_buildUrl[${this.name}]`;
+		logger.debug(`${method} - start`);
+
+		let t_hostname = hostname;
+		// endpoints may be running in containers on the local system
+		if (this.asLocalhost) {
+			t_hostname = 'localhost';
 		}
 
-		return false;
+		// If we connect to a discovery peer over TLS, any endpoints returned by
+		// discovery should also use TLS.
+		let protocol = null;
+		let isTLS = true;
+		if (this.currentTarget) {
+			isTLS = this.currentTarget.endpoint.isTLS();
+		}
+		protocol = isTLS ? 'grpcs' : 'grpc';
+		// but if not, use the following to override
+		const overrideProtocol = this.client.getConfigSetting('discovery-override-protocol');
+		if (overrideProtocol) {
+			protocol = overrideProtocol;
+		}
+
+		return `${protocol}://${t_hostname}:${port}`;
 	}
 
 	/* internal method
@@ -449,331 +450,6 @@ class DiscoveryService extends ServiceAction {
 		return chaincodeInterest;
 	}
 
-	// -- process the ChaincodeQueryResult - fabproto6.discovery.QueryResult.ChaincodeQueryResult
-	async _processChaincode(q_chaincodes) {
-		const method = '_processChaincode';
-		logger.debug(`${method} - start`);
-		const endorsement_plan = {};
-		// repeated EndorsementDescriptor content, but we should only have one
-		if (q_chaincodes && q_chaincodes.content && Array.isArray(q_chaincodes.content)) {
-			for (const q_endors_desc of q_chaincodes.content) {
-				endorsement_plan.chaincode = q_endors_desc.chaincode;
-
-				// named groups of Peers
-				endorsement_plan.groups = {};
-				for (const group_name in q_endors_desc.endorsers_by_groups) {
-					logger.debug(`${method} - found group: ${group_name}`);
-					const group = {};
-					group.peers = await this._processPeers(q_endors_desc.endorsers_by_groups[group_name].peers);
-					// all done with this group
-					endorsement_plan.groups[group_name] = group;
-				}
-
-				// LAYOUTS
-				endorsement_plan.layouts = [];
-				for (const q_layout of q_endors_desc.layouts) {
-					const layout = Object.assign({}, q_layout.quantities_by_group);
-					logger.debug(`${method} - layout :${layout}`);
-					endorsement_plan.layouts.push(layout);
-				}
-			}
-		} else {
-			throw Error('Plan layouts are invalid');
-		}
-
-		return endorsement_plan;
-	}
-
-	_processConfig(q_config) {
-		const method = `_processConfig[${this.name}]`;
-		logger.debug(`${method} - start`);
-		const config = {};
-		config.msps = {};
-		config.orderers = {};
-
-		try {
-			if (q_config.msps) {
-				for (const id in q_config.msps) {
-					logger.debug(`${method} - found organization ${id}`);
-					const q_msp = q_config.msps[id];
-					const msp_config = {
-						id: id,
-						name: id,
-						organizationalUnitIdentifiers: q_msp.organizational_unit_identifiers,
-						rootCerts: byteToNormalizedPEM(q_msp.root_certs),
-						intermediateCerts: byteToNormalizedPEM(q_msp.intermediate_certs),
-						admins: byteToNormalizedPEM(q_msp.admins),
-						tlsRootCerts: byteToNormalizedPEM(q_msp.tls_root_certs),
-						tlsIntermediateCerts: byteToNormalizedPEM(q_msp.tls_intermediate_certs)
-					};
-					config.msps[id] = msp_config;
-					this.channel.addMsp(msp_config, true);
-				}
-			} else {
-				logger.debug(`${method} - no msps found`);
-			}
-			/*
-			"orderers":{"OrdererMSP":{"endpoint":[{"host":"orderer.example.com","port":7050}]}}}
-			*/
-			if (q_config.orderers) {
-				for (const mspid in q_config.orderers) {
-					logger.debug(`${method} - found orderer org: ${mspid}`);
-					config.orderers[mspid] = {};
-					config.orderers[mspid].endpoints = [];
-					for (const endpoint of q_config.orderers[mspid].endpoint) {
-						config.orderers[mspid].endpoints.push(endpoint);
-					}
-				}
-			} else {
-				logger.debug(`${method} - no orderers found`);
-			}
-		} catch (err) {
-			logger.error(`${method} - Problem with discovery config: ${err}`);
-		}
-
-		return config;
-	}
-
-	async _processMembership(q_members) {
-		const method = `_processMembership[${this.name}]`;
-		logger.debug(`${method} - start`);
-		const peersByOrg = {};
-		if (q_members.peers_by_org) {
-			for (const mspid in q_members.peers_by_org) {
-				logger.debug(`${method} - found org:${mspid}`);
-				peersByOrg[mspid] = {};
-				peersByOrg[mspid].peers = await this._processPeers(q_members.peers_by_org[mspid].peers);
-			}
-		} else {
-			logger.debug(`${method} - missing peers by org`);
-		}
-		return peersByOrg;
-	}
-
-	// message Peers
-	async _processPeers(q_peers) {
-		const method = `_processPeers[${this.name}]`;
-		const peers = [];
-		// message Peer
-		for (const q_peer of q_peers) {
-			const peer = {};
-			// IDENTITY
-			const q_identity = fabproto6.msp.SerializedIdentity.decode(q_peer.identity);
-			peer.mspid = q_identity.mspid;
-
-			// MEMBERSHIP - Peer.membership_info
-			// fabproto6.gossip.Envelope.payload
-			const q_membership_message = fabproto6.gossip.GossipMessage.decode(q_peer.membership_info.payload);
-			peer.endpoint = q_membership_message.alive_msg.membership.endpoint;
-			peer.name = q_membership_message.alive_msg.membership.endpoint;
-			logger.debug(`${method} - peer :${peer.endpoint}`);
-
-			// STATE
-			if (q_peer.state_info) {
-				const message_s = fabproto6.gossip.GossipMessage.decode(q_peer.state_info.payload);
-				peer.ledgerHeight = Long.fromValue(message_s.state_info.properties.ledger_height);
-				logger.debug(`${method} - ledgerHeight :${peer.ledgerHeight}`);
-				peer.chaincodes = [];
-				for (const index in message_s.state_info.properties.chaincodes) {
-					const q_chaincode = message_s.state_info.properties.chaincodes[index];
-					const chaincode = {};
-					chaincode.name = q_chaincode.name;
-					chaincode.version = q_chaincode.version;
-					// TODO metadata ?
-					logger.debug(`${method} - chaincode :${JSON.stringify(chaincode)}`);
-					peer.chaincodes.push(chaincode);
-				}
-			} else {
-				logger.debug(`${method} - no state info for peer ${peer.endpoint}`);
-			}
-
-			// all done with this peer
-			peers.push(peer);
-			// build the GRPC instance
-			await this._buildPeer(peer);
-		}
-
-		return peers;
-	}
-
-	async _buildOrderers(orderers) {
-		const method = `_buildOrderers[${this.name}]`;
-		logger.debug(`${method} - start`);
-
-		if (!orderers) {
-			logger.debug('%s - no orderers to build', method);
-		} else {
-			for (const msp_id in orderers) {
-				logger.debug(`${method} - orderer msp:${msp_id}`);
-				for (const endpoint of orderers[msp_id].endpoints) {
-					endpoint.name = await this._buildOrderer(endpoint.host, endpoint.port, msp_id);
-				}
-			}
-		}
-
-		return orderers;
-	}
-
-	async _buildOrderer(host, port, msp_id) {
-		const method = `_buildOrderer[${this.name}]`;
-		logger.debug(`${method} - start mspid:${msp_id} endpoint:${host}:${port}`);
-
-		const name = `${host}:${port}`;
-		const url = this._buildUrl(host, port);
-		logger.debug(`${method} - create a new orderer ${url}`);
-		const orderer = this.client.newCommitter(name, msp_id);
-		const end_point = this.client.newEndpoint(this._buildOptions(name, url, host, msp_id));
-		try {
-			// first check to see if orderer is already on this channel
-			let same;
-			const channelOrderers = this.channel.getCommitters();
-			for (const channelOrderer of channelOrderers) {
-				logger.debug('%s - checking %s', method, channelOrderer);
-				if (channelOrderer.endpoint && channelOrderer.endpoint.url === url) {
-					same = channelOrderer;
-					break;
-				}
-			}
-			if (!same) {
-				await orderer.connect(end_point);
-				this.channel.addCommitter(orderer);
-			} else {
-				await same.checkConnection();
-				logger.debug('%s - orderer already added to this channel', method);
-			}
-		} catch (error) {
-			logger.error(`${method} - Unable to connect to the discovered orderer ${name} due to ${error}`);
-		}
-
-		return name;
-	}
-
-	async _buildPeer(discovery_peer) {
-		const method = `_buildPeer[${this.name}]`;
-		logger.debug(`${method} - start`);
-
-		if (!discovery_peer) {
-			throw Error('Missing discovery_peer parameter');
-		}
-		const address = discovery_peer.endpoint;
-		const msp_id = discovery_peer.mspid;
-
-		const host_port = address.split(':');
-		const url = this._buildUrl(host_port[0], host_port[1]);
-
-		// first check to see if peer is already on this channel
-		let peer;
-		const channelPeers = this.channel.getEndorsers();
-		for (const channelPeer of channelPeers) {
-			logger.debug('%s - checking channel peer %s', method, channelPeer.name);
-			if (channelPeer.endpoint && channelPeer.endpoint.url === url) {
-				logger.debug('%s - url: %s - already added to this channel', method, url);
-				peer = channelPeer;
-				break;
-			}
-		}
-		if (!peer) {
-			logger.debug(`${method} - create a new endorser ${url}`);
-			peer = this.client.newEndorser(address, msp_id);
-			const end_point = this.client.newEndpoint(this._buildOptions(address, url, host_port[0], msp_id));
-			try {
-				logger.debug(`${method} - about to connect to endorser ${address} url:${url}`);
-				await peer.connect(end_point);
-				this.channel.addEndorser(peer);
-				logger.debug(`${method} - connected to peer ${address} url:${url}`);
-			} catch (error) {
-				logger.error(`${method} - Unable to connect to the discovered peer ${address} due to ${error}`);
-			}
-		} else {
-			// make sure the existing connect is still good
-			await peer.checkConnection();
-		}
-
-		// indicate that this peer has been touched by the discovery service
-		peer.discovered = true;
-
-		// make sure that this peer has all the found installed chaincodes
-		if (discovery_peer.chaincodes) {
-			for (const chaincode of discovery_peer.chaincodes) {
-				logger.debug(`${method} - adding chaincode ${chaincode.name} to peer ${peer.name}`);
-				peer.addChaincode(chaincode.name);
-			}
-		}
-
-		logger.debug(`${method} - end`);
-		return peer;
-	}
-
-	_buildUrl(hostname = checkParameter('hostname'), port = checkParameter('port')) {
-		const method = `_buildUrl[${this.name}]`;
-		logger.debug(`${method} - start`);
-
-		let t_hostname = hostname;
-		// endpoints may be running in containers on the local system
-		if (this.asLocalhost) {
-			t_hostname = 'localhost';
-		}
-
-		// If we connect to a discovery peer over TLS, any endpoints returned by
-		// discovery should also use TLS.
-		let protocol = null;
-		let isTLS = true;
-		if (this.currentTarget) {
-			isTLS = this.currentTarget.endpoint.isTLS();
-		}
-		protocol = isTLS ? 'grpcs' : 'grpc';
-		// but if not, use the following to override
-		const overrideProtocol = this.client.getConfigSetting('discovery-override-protocol');
-		if (overrideProtocol) {
-			protocol = overrideProtocol;
-		}
-
-		return `${protocol}://${t_hostname}:${port}`;
-	}
-
-	_buildOptions(name, url, host, msp_id) {
-		const method = `_buildOptions[${this.name}]`;
-		logger.debug(`${method} - start`);
-		const caroots = this._buildTlsRootCerts(msp_id);
-		return {
-			url: url,
-			pem: caroots,
-			'ssl-target-name-override': host,
-			name: name
-		};
-	}
-
-	_buildTlsRootCerts(msp_id = checkParameter('msp_id')) {
-		const method = `_buildTlsRootCerts[${this.name}]`;
-		logger.debug(`${method} - start`);
-		let ca_roots = '';
-		let mspDiscovered;
-		if (this.discoveryResults && this.discoveryResults.msps) {
-			mspDiscovered = this.discoveryResults.msps[msp_id];
-		} else {
-			logger.error('Missing MSPs discovery results');
-			return ca_roots;
-		}
-		if (mspDiscovered) {
-			logger.debug(`Found msp ${msp_id}`);
-		} else {
-			logger.error(`Missing msp ${msp_id} in discovery results`);
-			return ca_roots;
-		}
-		if (mspDiscovered.tlsRootCerts) {
-			ca_roots = ca_roots + mspDiscovered.tlsRootCerts;
-		} else {
-			logger.debug('%s - no tls root certs', method);
-		}
-		if (mspDiscovered.tlsIntermediateCerts) {
-			ca_roots = ca_roots + mspDiscovered.tlsIntermediateCerts;
-		} else {
-			logger.debug('%s - no tls intermediate certs', method);
-		}
-
-		return ca_roots;
-	}
-
 	/**
 	 * Close the connection of the discovery service.
 	 */
@@ -792,7 +468,6 @@ class DiscoveryService extends ServiceAction {
 	 * return a printable representation of this object
 	 */
 	toString() {
-
 		return `DiscoveryService: {name: ${this.name}, channel: ${this.channel.name}}`;
 	}
 }
@@ -813,6 +488,10 @@ function _getCollectionNames(names, chaincodeCall) {
 	} else {
 		throw Error('Collection names must be an array of strings');
 	}
+}
+
+function currentTimestamp() {
+	return new Date().getTime();
 }
 
 module.exports = DiscoveryService;
